the glass is not gone after setting 0->x->0
with this change it is possible to get a maximized window with hidden
taskbar.
so we can set initial ```GlassFrameThickness = new Thickness(0)``` with
```WindowStyle=None``` to get a custom window.
now if user maximize the window we can set ```GlassFrameThickness = new
Thickness(1)``` to get a window that fills the complete monitor and
hides the taskbar.
with ```GlassFrameThickness = new Thickness(0)``` we can set back to a
custom styled window without showing the glass frame.
this all works only with ```WindowStyle=None```!