fix disappearing taskbar item
Windows 8.1 Settings:

2 monitors with extended desktop
Taskbar only visible on primary desktop
Repro:

Start the WPF application
Move the window on the secondary screen
Maximize the window on the secondary screen (for example by docking the
window on the top)
Restore and drag the window from the secondary screen to the primary
screen
--> The taskbar icon will disappear exactly when the mouse enters on the
primary screen!

If you do the same repro again the icon reappears.

Another repro:

- Maximize the application on second monitor
- Minimize the app and focus another application OR click somewhere on a
second monitor
- Focus/Activate the original application again
- The taskbar item disappeared (maybe only after second repro)